Move routes to their own file
